refactor: create SearchInput component from Search
